Added world plugin for ldtk
