+++ src/lib.rs
@@ -15,11 +15,15 @@ mod loading;
 mod menu;
 mod player;
 
+mod world;
+
 use crate::actions::ActionsPlugin;
 use crate::audio::InternalAudioPlugin;
 use crate::loading::LoadingPlugin;
 use crate::menu::MenuPlugin;
 use crate::player::PlayerPlugin;
+
+use crate::world::WorldPlugin;
 
 use bevy::app::App;
 #[cfg(debug_assertions)]
@@ -50,6 +54,8 @@ impl Plugin for GamePlugin {
             ActionsPlugin,
             // InternalAudioPlugin,
             // PlayerPlugin,
+
+            WorldPlugin,
         ));
 
         #[cfg(debug_assertions)]
